:gem: :racehorse: :fu: Add a couple of new queries for #17. Improve performance slightly for others, and refactor lists to be like the rest of the query categories.
