+++ lib/mediawiki/query/lists/lists.rb
@@ -1,0 +1,19 @@
+require_relative '../../constants'
+require_relative '../query'
+require_relative 'backlinks'
+require_relative 'categories'
+require_relative 'all'
+require_relative 'search'
+require_relative 'miscellaneous'
+
+module MediaWiki
+  module Query
+    module Lists
+      include MediaWiki::Query::Lists::Backlinks
+      include MediaWiki::Query::Lists::Categories
+      include MediaWiki::Query::Lists::All
+      include MediaWiki::Query::Lists::Search
+      include MediaWiki::Query::Lists::Miscellaneous
+    end
+  end
+end
